fix: CGO type conversion on Windows
- Fix C.size_t() conversion to use uintptr() intermediate cast
- Resolves Windows compilation errors with uint32 vs size_t mismatch
- Affects: table.go and connection.go IPC methods

+++ pkg/internal/connection.go
@@ -190,7 +190,7 @@ func (c *Connection) CreateTable(ctx context.Context, name string, schema contra
 		c.handle,
 		cName,
 		cSchemaPtr,
-		C.size_t(len(schemaIPC)),
+		C.size_t(uintptr(len(schemaIPC))),
 	)
 	defer C.simple_lancedb_result_free(result)
